src/app/+targets/shared/+targets-details/targets-details.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { TargetsDetailsComponent } from './targets-details.component';

describe('Component: Details', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [TargetsDetailsComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([TargetsDetailsComponent],
    (component: TargetsDetailsComponent) => {
      expect(component).toBeTruthy();
    }));

  it('should create the component', inject([], () => {
    return builder.createAsync(TargetsDetailsTestComponent)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(TargetsDetailsComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-details></app-details>
  `,
  directives: [TargetsDetailsComponent]
})
class TargetsDetailsTestComponent {
}

